+++ src/server/service/batch.rs
@@ -10,7 +10,7 @@ use crate::server::service::kv::{
     batch_commands_request, batch_commands_response, future_batch_get_command,
     future_raw_batch_get_command, poll_future_notify,
 };
-use crate::storage::{kv::Engine, lock_manager::LockManager, txn::PointGetCommand, Storage};
+use crate::storage::{kv::Engine, lock_manager::LockManager, PointGetCommand, Storage};
 use kvproto::kvrpcpb::*;
 use tikv_util::collections::HashMap;
 use tikv_util::metrics::HistogramReader;
